LayoutTests/crypto/subtle/jwk-export-use-values-expected.txt

Test exporting keys with various usages to JWK.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


encrypt:
PASS jwk.use is undefined
PASS jwk.key_ops is ["encrypt"]

decrypt:
PASS jwk.use is undefined
PASS jwk.key_ops is ["decrypt"]

encrypt,decrypt:
PASS jwk.use is undefined
PASS jwk.key_ops is ["encrypt","decrypt"]

wrapKey:
PASS jwk.use is undefined
PASS jwk.key_ops is ["wrap"]

unwrapKey:
PASS jwk.use is undefined
PASS jwk.key_ops is ["unwrap"]

wrapKey,unwrapKey:
PASS jwk.use is undefined
PASS jwk.key_ops is ["wrap","unwrap"]

encrypt,decrypt,wrapKey:
PASS jwk.use is undefined
PASS jwk.key_ops is ["encrypt","decrypt","wrap"]

encrypt,decrypt,wrapKey,unwrapKey:
PASS jwk.use is undefined
PASS jwk.key_ops is ["encrypt","decrypt","wrap","unwrap"]

sign:
PASS jwk.use is undefined
PASS jwk.key_ops is ["sign"]

verify:
PASS jwk.use is undefined
PASS jwk.key_ops is ["verify"]

sign,verify:
PASS jwk.use is undefined
PASS jwk.key_ops is ["sign","verify"]

PASS successfullyParsed is true

TEST COMPLETE
